They changed the template. I added the changes.
