+++ README.md
@@ -2,7 +2,9 @@ This project was bootstrapped with [Create React App](https://github.com/faceboo
 This project contains an example project board meant to showcase how one can be used. The issues posted to it are not real issues.
 
 #### _**IMPORTANT NOTE**_ - 
-This project does not have a mongoDB connection setup. To set this up first uncomment the code at the top of server.js and then create a config file which exports your db.uri connection.
+This project does not have a mongoDB connection setup. For:
+- local development: create a config file (make sure to name it config.js) in the config folder, which exports your db.uri connection. An example is provided, config/config.example.js. This file will be ignored by git so your db credentials will be kept safe when the app is deployed.
+- production: Since the config file is not pushed when you deploy your app, you must specifiy your db uri in heorku. Set the uri in heroku as specified in [this](https://devcenter.heroku.com/articles/config-vars) resource. Make sure you name the environement variable "DB_URI".
 
 ## Getting Started
 This repository aims to assist you in beginning work on a MERN stack application with a solid file structure as a foundation. To get started make a copy of this template repo for your project teams.
